Updated build.gradle
Use spotless
