+++ gradle/libs.versions.toml
@@ -1,0 +1,15 @@
+[versions]
+log4j = "2.25.1"
+
+[plugins]
+spotless = { id = "com.diffplug.spotless", version = "7.2.1" }
+versions = { id = "com.github.ben-manes.versions", version = "0.52.0" }
+
+[libraries]
+jetbrains-annotations = { module = "org.jetbrains:annotations", version = "26.0.2" }
+junit-jupiter = { module = "org.junit.jupiter:junit-jupiter", version = "5.13.4" }
+junit-platform = { module = "org.junit.platform:junit-platform-launcher", version = "1.13.4" }
+log4j-api = { module = "org.apache.logging.log4j:log4j-api", version.ref = "log4j" }
+log4j-core = { module = "org.apache.logging.log4j:log4j-core", version.ref = "log4j" }
+
+[bundles]
