+++ settings.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+  id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+}
+
+rootProject.name = "Keschet"
