--- src/main/java/github/buried_in_code/Direction.java
@@ -1,15 +1,0 @@
-package github.buried_in_code;
-
-/**
- * Created by BuriedInCode on 2018-Feb-08.
- */
-public enum Direction {
-	NORTH,
-	NORTH_EAST,
-	EAST,
-	SOUTH_EAST,
-	SOUTH,
-	SOUTH_WEST,
-	WEST,
-	NORTH_WEST
-}
