--- src/main/java/github/buried_in_code/pieces/General.java
@@ -1,30 +1,0 @@
-package github.buried_in_code.pieces;
-
-import github.buried_in_code.Direction;
-import github.buried_in_code.players.Player;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-/**
- * Created by BuriedInCode on 2018-Feb-09.
- */
-public class General extends Piece {
-	public General(@NotNull Player player) {
-		super(player, 10, "G", List.of(
-				Direction.NORTH,
-				Direction.NORTH_EAST,
-				Direction.EAST,
-				Direction.SOUTH_EAST,
-				Direction.SOUTH,
-				Direction.SOUTH_WEST,
-				Direction.WEST,
-				Direction.NORTH_WEST
-		));
-	}
-
-	@Override
-	public String toString() {
-		return "General{} " + super.toString();
-	}
-}
