+++ src/main/java/github/buriedincode/Direction.java
@@ -1,0 +1,5 @@
+package github.buriedincode;
+
+public enum Direction {
+  NORTH, NORTH_EAST, EAST, SOUTH_EAST, SOUTH, SOUTH_WEST, WEST, NORTH_WEST
+}
